feat: Implement global exception handling for validation and runtime errors

+++ Backend/OrdersServices/src/main/java/com/EComMicroService/OrdersServices/KafkaEvents/Events.java
@@ -20,8 +20,8 @@ public class Events {
     private String orderNumber;
     private EventType eventType;
     private Map<String,Integer> items;
-    private long discountAmount;
-    private long total;
+    private float discountAmount;
+    private float total;
     private Address shippingAddress;
 }
